requirements.txt
pygame
Pillow
google-generativeai
python-dotenv
